Refactor Bulk Editor Content to Use Inline Editing Hook
